LPS-76116 Minor change to move the converter to internal package
