+++ modules/apps/web-experience/export-import/export-import-service/src/main/resources/com/liferay/exportimport/xstream/packageinfo
@@ -1,1 +1,1 @@
-version 1.2.0
+version 1.1.1
